expressions a la place de condition

+++ PLD-COMP-RENDU/pld-comp/compiler/symbolTable.cpp
@@ -13,7 +13,7 @@
 	{
 		size_t index =0;
 		if(typeStr=="int"){
-			index = sizeof(int)*(symbols.size()+1);
+			index = sizeof(int64_t)*(symbols.size()+1);
 		}else if(typeStr=="char"){
 			index = sizeof(char)*(symbols.size()+1);
 		}
